stock-backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
pymysql
python-dotenv
